src/requirements.txt
#!python3.12
pyvsc
ruamel.yaml
pyelftools
